sync stuff

+++ src/main/java/pepjebs/mapatlases/client/screen/MapWidget.java
@@ -95,7 +95,7 @@ public class MapWidget extends AbstractAtlasWidget implements Renderable, GuiEve
         this.isHovered = isMouseOver(pMouseX, pMouseY);
 
         // Handle zooming markers hack
-        MapAtlasesClient.setDecorationsScale(zoomLevel * (float) (double) MapAtlasesClientConfig.worldMapDecorationScale.get());
+        MapAtlasesClient.setDecorationsScale( zoomLevel * (float)(double) MapAtlasesClientConfig.worldMapDecorationScale.get());
 
         this.drawAtlas(graphics, x, y, width, height, player, zoomLevel,
                 MapAtlasesClientConfig.worldMapBorder.get());
@@ -277,7 +277,7 @@ public class MapWidget extends AbstractAtlasWidget implements Renderable, GuiEve
             animationProgress = Math.max(0, animationProgress);
         }
         if (this.zoomLevel != targetZoomLevel) {
-            zoomLevel = interpolate(targetZoomLevel, zoomLevel, animationSpeed);
+            zoomLevel = (float) interpolate(targetZoomLevel, zoomLevel, animationSpeed);
         }
         boolean test = true;
         if (this.currentXCenter != targetXCenter) {
@@ -299,12 +299,12 @@ public class MapWidget extends AbstractAtlasWidget implements Renderable, GuiEve
         }
     }
 
-    private float interpolate(float targetZCenter, float currentZCenter, float animationSpeed) {
-        float diff = targetZCenter - currentZCenter;
+    private double interpolate(double targetZCenter, double currentZCenter, double animationSpeed) {
+        double diff = targetZCenter - currentZCenter;
         if (diff < 0) {
-            return Math.max(targetZCenter, currentZCenter + (diff * animationSpeed) - 0.001f);
+            return Math.max(targetZCenter, currentZCenter + (diff * animationSpeed) - 0.001);
         } else {
-            return Math.min(targetZCenter, currentZCenter + (diff * animationSpeed) + 0.001f);
+            return Math.min(targetZCenter, currentZCenter + (diff * animationSpeed) + 0.001);
         }
     }
 }
